requirements.txt
sentence-transformers>=2.2.2
qdrant-client>=1.8.0
streamlit>=1.34.0
datasets>=2.14.0
pillow>=10.0.0
tqdm>=4.65.0
python-dotenv>=0.19.0
torch>=2.0.0